requirements.txt
streamlit==1.32.0
python-dotenv==1.0.1
google-generativeai==0.3.2
requests
streamlit-chat==0.1.1
